+++ tests/testthat/test-tree_comparison.R
@@ -1,0 +1,13 @@
+test_that('Pairwise distances calculated correctly', {
+  nTrees <- 6L
+  nTip <- 16L
+
+  set.seed(0)
+  trees <- lapply(rep(nTip, nTrees), RandomTree, root = TRUE)
+  trees[[1]] <- BalancedTree(nTip)
+  trees[[nTrees - 1L]] <- PectinateTree(nTip)
+  class(trees) <- 'multiPhylo'
+
+  dists <- PairwiseDistances(trees, phangorn::RF.dist)
+  expect_equivalent(phangorn::RF.dist(trees), dists)
+})